fast_socks5::client

Struct Socks5Stream

pub struct Socks5Stream<S: AsyncRead + AsyncWrite + Unpin> { /* private fields */ }

A SOCKS5 client. Socks5Stream implements [AsyncRead] and [AsyncWrite].

Implementations

impl<S> Socks5Stream<S>
where S: AsyncRead + AsyncWrite + Unpin,

pub async fn use_stream( socket: S, auth: Option<AuthenticationMethod>, config: Config, ) -> Result<Self>

Possibility to use a stream already created rather than creating a whole new TcpStream::connect().

pub async fn request( &mut self, cmd: Socks5Command, target_addr: TargetAddr, ) -> Result<TargetAddr>

pub fn get_socket(self) -> S

pub fn get_socket_ref(&self) -> &S

pub fn get_socket_mut(&mut self) -> &mut S

impl Socks5Stream<TcpStream>

Api if you want to use TcpStream to create a new connection to the SOCKS5 server.

pub async fn connect<T>( socks_server: T, target_addr: String, target_port: u16, config: Config, ) -> Result<Self>
where T: ToSocketAddrs,

Connects to a target server through a SOCKS5 proxy.

pub async fn connect_with_password<T>( socks_server: T, target_addr: String, target_port: u16, username: String, password: String, config: Config, ) -> Result<Self>
where T: ToSocketAddrs,

Connect with credentials

pub async fn connect_raw<T>( cmd: Socks5Command, socks_server: T, target_addr: String, target_port: u16, auth: Option<AuthenticationMethod>, config: Config, ) -> Result<Self>
where T: ToSocketAddrs,

Process clients SOCKS requests This is the entry point where a whole request is processed.