fast_socks5

Module server

Structs

AcceptAuthentication

While this one will always allow the user in.

AuthSucceeded

The struct returned when the user has successfully authenticated

Config

DenyAuthentication

This will simply return Option::None, which denies the authentication

Incoming

Incoming implements [futures::stream::Stream].

SimpleUserPassword

Basic user/pass auth method provided.

Socks5Server

Wrapper of TcpListener Useful if you don’t use any existing TcpListener’s streams.

Socks5Socket

Wrap TcpStream and contains Socks5 protocol implementation.

Traits

Authentication

Use this trait to handle a custom authentication on your end.